code review comments addressed #5
