+++ tools/content-loader/src/helpers/helm-tpl-value-resolver.unit.test.ts
@@ -1,4 +1,4 @@
-import { parentDir, valueTplMatchGenerator, merge } from "./value-resolver";
+import { valueTplMatchGenerator, merge } from "./helm-tpl-value-resolver";
 import to from "await-to-js";
 
 describe("merge", () => {
@@ -43,27 +43,6 @@ describe("merge", () => {
       expect(actual).toEqual(expected);
     },
   );
-});
-
-describe("parentDir", () => {
-  const cases = [
-    [
-      "/test/me", // path
-      "/test", // expected
-    ],
-    [
-      "/Applications/Visual Studio Code - Insiders.app/Contents/Frameworks", // path
-      "/Applications/Visual Studio Code - Insiders.app/Contents", // expected
-    ],
-    [
-      "/Applications/Visual Studio Code - Insiders.app", // path
-      "/Applications", // expected
-    ],
-  ];
-  test.each(cases)('it should resolve "%s" to "%s"', (path, expected) => {
-    const actual = parentDir(path);
-    expect(actual).toEqual(expected);
-  });
 });
 
 describe("template regexp", () => {
