+++ tools/content-loader/src/helpers/parent-dir.ts
@@ -1,0 +1,9 @@
+import { sep } from "path";
+
+export default (p: string) => {
+  if (!p) {
+    return "";
+  }
+  const parent = p.substring(0, p.lastIndexOf(sep));
+  return parent ? parent : "";
+};
